Fix mounting custom directories via sshfs
It appears that the script was attempting to allow URLs with a colon
present to pass through unmodified, so someone could specify a custom
directory to found, such as `ssh://host.example.com:/path/on/remote`.
However, the regex check for  URLs was testing for the character
literals :" (colon, double quote) within a protocol-stripped URL (which
it has been at this point int he code, being
`host.example.com:/path/on/remote`). Given how the later `sshfs` line
operates (wraps $FINAL_REMOTE_URL in double quotes), if one was to
specify a URL with `:"`, then it'll never end up working when passed to
sshfs.

This change simplifies the regex match for stripped URLs to just a `.*:`
(colon character somewhere) such that if schapps detects a URL that's
been passed to it like the first example above, it will mount
`/path/on/remote` from `host.example.com`.

Prior to this patch, there was no way of setting `remote.url` to a value
that would allow mounting a specific path over sshfs; one could only
mount `host.example.com:` (the default, which corresponds to the user's
home directory).

Given how the original code was written, it seems like it might have
just been a typo having the double-quote character there.