Revert "Add support for certbackup"
This reverts commit 7ad0ba8301f61a465a4668915c1198149278bd5f.

Certificate backup restore is handled by hook run by schnapps so this
code duplicates that functionality. It should not be there.

It is also wrong as some of the files are hidden (starting with dot)
and this copy omits them. At the same time it prints error if there is
no file in store (this is "fixed" by hiding errors but nooo).