Move the rest of configuration parsing back
Turns out that we are actually using the parsed configuration in showing
help, so let's keep all the configuration together. It is nicer that way
anyway. But let's move the die function to the top so we can use it
during configuration parsing. Fortunately we don't need die_hepling
function which would cause a cycle.