Allow exporting to stdout
Exporting to stdout allows us to chain exported output to additional
commands without storing files temporally somewhere.

The disadvantage of such processing is that there is not info file and
thus can't be imported later on without creating one. It can be used
only to investigate content (such as in support workflow) or as the
medkit.