Support local:// URI for "remote" storage
We had a request to backup snapshots to locally mounted drive. It makes
sense, if you have for example connected external drive. In theory I
could have switched to the directory and work there, but wanted to
simplify working with remote storages and make it as similar as possible
so I'm using bind mount. Should nicely "isolate" and test the selected
storage.