Adapt to new quota output
Showing size of the snapshots no longer works. In the past the output of
the quota command looked like this:

qgroupid         rfer         excl
--------         ----         ----
0/5          16.00KiB     16.00KiB

Nowadays it contains a patch as well and the final output is as follows:

Qgroupid    Referenced    Exclusive   Path
--------    ----------    ---------   ----
0/1386         2.46GiB     60.34MiB   @135

So updating schnapps to be able to parse the second output as well. In
theory even better would be to use json output and do the math to get
human readable output, but doing math in shell is not a great idea, so
sticking with the text output for now.